refactor: nomenclatura de stores
